+++ CHANGELOG.rst
@@ -21,7 +21,8 @@ Added
     - IP address based access restriction for projects (#531)
     - ``is_delegate()`` and ``is_owner_or_delegate()`` helpers for ``Project`` model
     - Remote sync for non-owner category members (#502)
-    - Added ``setting_delete()`` function to ``AppSettingAPI`` (#538)
+    - ``setting_delete()`` function to ``AppSettingAPI`` (#538)
+    - ``cleanappsettings`` management command (#374)
 
 Changed
 -------